refactor: move interfaces into seperate folder
