+++ src/app/video/pipes/filter-videos-by-title.pipe.ts
@@ -1,5 +1,5 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { Video } from '../../models/video';
+import { Video } from '../../models/video.interface';
 
 @Pipe({
   name: 'filterVideosByTitle',
