+++ src/app/video/video-item/video-item.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { Video } from '../../models/video';
+import { Video } from '../../models/video.interface';
 
 @Component({
   selector: 'app-video-item',
